fix: route authenticated 404 users home to notes (#347)
